Write a function called alter_list. alter_list should have two parameters: a list of strings and a list of integers.
The list of integers will represent indices for the list of
strings. alter_list should alter the capitalization of all
the words at the designated indices. If the word was all
capitals, it should become all lower case. If it was all
lower case, it should become all capitals. You may assume
that the words will already be all-caps or all-lower case.